Fixes bug with inviting friends to trip
Inviting friends to trip had no gaurd against submitting an email
that belonged to a registered user. Any emails belonging to a registered
user are converted and added as an id, unless they are already invited
or active on a trip, in which case, the id is purged from the invite
process.

Adds more tests for inviting friends.

Adds tests for signup process.

Convert TripUser model to use eloquent realtions instead of accessing
directly as a model.

Removes unused RequestController.